requirements.txt
langchain-openai==0.0.2
python-dotenv==1.0.0
streamlit==1.29.0
